% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{Prices of over 50,000 round cut diamonds}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{name}{name}
\item{date}{random date}
\item{payment}{amount of payment}
}
}
\usage{
df
}
\description{
A dataset containing random data. The variables are as follows:
}
\keyword{datasets}
