% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_format.R
\name{date_format}
\alias{date_format}
\title{Transforming Date column}
\usage{
date_format(df, date)
}
\arguments{
\item{df}{The data table}

\item{date}{The column indicating the date}
}
\description{
Following function transforms the Date as class Date and creates a new
columns with the Date as Month and Year. This is especially handy for
the creation of visualization as it takes less space and avoids overlapping
text on the axes.
}
\examples{
date_format(df, date)
}
