% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{df}
\alias{df}
\title{Random data of clients}
\format{
A data frame with random variables:
\describe{
\item{name}{name}
\item{date}{random date}
\item{payment}{amount of payment}
}
}
\description{
A dataset containing random data. The variables are as follows:
}
