% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst_pkg.R
\name{inst_pkg}
\alias{inst_pkg}
\title{Install and load packages}
\usage{
inst_pkg(pkg)
}
\arguments{
\item{pkg}{Include the packages you like to install}
}
\description{
This function allows you to install and load all your packages. Additionally
it is shown whether the package was loaded successfully or not.
}
\examples{
pkg = c("ggplot2")
inst_pkg(pkg)
}
